Working on partial/delta update support Getting mouse control working Starting to fix game/server time

+++ UdpHosts/MyGameServer/AIEngine.cs
@@ -8,7 +8,7 @@ namespace MyGameServer {
 
 		}
 
-		public void Tick( double deltaTime, double currTime ) {
+		public void Tick( double deltaTime, uint currTime ) {
 
 		}
 	}
